Create a new sequnce.js file and move existing functions
